Comment all subroutines so that execution path can be traced.
This version works in a minimal way with moodle 1.9.  The connection
for returning the answers still needs work.  The way the variables is
being passed doesn't quite match up.  I believe the answers are
correctly reaching the server but their may be issues on the server
side with processing them and there are definitely issues with
passing the answers and the resulting grade back.

This is a reasonable marker in the process of getting these files to
work again.  Still needs much work.